Move password check logic into AccountManagementService
